refactor(harness): drop host.events compatibility shim
Remove the 3-line re-export shim host/events.py and import
EventListener/OrderedEventBus directly from loushang.harness.events.
This breaks the host <-> events circular edge at the package boundary.
