--- src/loushang/harness/host/events.py
@@ -1,5 +1,0 @@
-"""Compatibility exports for the shared runtime event bus."""
-
-from loushang.harness.events import EventListener, OrderedEventBus
-
-__all__ = ["EventListener", "OrderedEventBus"]
